Move example test classes to dedicated subproject
