+++ library/build.gradle
@@ -83,8 +83,7 @@ dependencies {
     compile('org.jetbrains.kotlin:kotlin-reflect')
     compile("org.junit.jupiter:junit-jupiter-api:5.1.0")
     compile("org.junit.platform:junit-platform-launcher:1.1.1")
-    compile('org.springframework.boot:spring-boot-starter-test') { exclude group: 'junit', module: 'junit' }
-    compile('org.springframework.boot:spring-boot-starter-web')
+    compile('org.springframework.boot:spring-boot-starter')
 
     // You can easily generate your own configuration metadata file from items annotated with
     // @ConfigurationProperties by using the spring-boot-configuration-processor jar. The jar
